+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+# 1.19.4
+* Get smooch apps from DynamoDB table.
+
 # 1.19.3
-* Add error info in log.error in set-conversation-retention-seconds
+* Add error info in log.error in set-conversation-retention-seconds.
 
 # 1.19.2
 * CXV1-21652 - Setting CX credentials for Register Gateway Lambda.